fix(gui-app): satisfy lint on worktree listing/enrichment/task-delete hooks
